move zipwith to tensor
