+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/tensor/DenseTensorApply.scala
@@ -275,4 +275,109 @@ object DenseTensorApply {
 
     (false, _offset)
   }
+
+  /**
+   * Iterate through tensor1, tensor2, and apply func to the elements,
+   * set function result to tensor 3
+   *
+   * @param tensor1 the tensor1
+   * @param tensor2 the tensor2
+   * @param tensor3 the result tensor
+   * @param func    (tensor1Data, tensor1Offset, tensor2Data,
+   *                tensor2Offset, tensor3Data, tensor3Offset)
+   */
+  def apply2[A, B, C](tensor1: Tensor[A], tensor2: Tensor[B], tensor3: Tensor[C],
+    func: (Array[A], Int, Array[B], Int, Array[C], Int) => Unit)
+  : Unit = {
+    require(tensor1.nElement() == tensor2.nElement(),
+      s"inconsistent tensor size: ${tensor1.nElement()} == ${tensor2.nElement()}")
+
+    if (tensor1.nDimension == 0) {
+      return
+    }
+
+    val tensor1Data = tensor1.storage().array()
+    var tensor1Offset = tensor1.storageOffset() - 1
+    val tensor2Data = tensor2.storage().array()
+    var tensor2Offset = tensor2.storageOffset() - 1
+    val tensor3Data = tensor3.storage().array()
+    val tensor3Offset = tensor3.storageOffset() - 1
+
+    var adjacent = false
+    if (tensor1.nDimension == 1 && tensor2.nDimension == 1 && tensor1.stride(1) == 1 &&
+      tensor2.stride(1) == 1) {
+      adjacent = true
+    }
+    if (tensor1.nDimension == 2 && tensor2.nDimension == 2) {
+      if (tensor1.stride(2) == 1 && tensor2.stride(2) == 1 && tensor1.stride(1) == tensor1.size(2)
+        && tensor2.stride(1) == tensor2.size(2)) {
+        adjacent = true
+      }
+
+      if (tensor1.stride(1) == 1 && tensor2.stride(1) == 1 && tensor1.stride(2) == tensor1.size(1)
+        && tensor2.stride(2) == tensor2.size(1)) {
+        adjacent = true
+      }
+    }
+    if (adjacent) {
+      var i = 0
+      while (i < tensor1.nElement()) {
+        func(
+          tensor1Data, tensor1Offset + i,
+          tensor2Data, tensor2Offset + i,
+          tensor3Data, tensor3Offset + i)
+        i += 1
+      }
+      return
+    }
+
+    val tensor1Stride = getStride(tensor1)
+    val (largestDim1, largestSize1) = getLargestContiguousSize(tensor1)
+    val counter1 = getCounter(largestDim1)
+    val tensor2Stride = getStride(tensor2)
+    val (largestDim2, largestSize2) = getLargestContiguousSize(tensor2)
+    val counter2 = getCounter(largestDim2)
+    val tensor3Stride = getStride(tensor3)
+    val (largestDim3, largestSize3) = getLargestContiguousSize(tensor3)
+    val counter3 = getCounter(largestDim3)
+
+    var hasFinished = false
+    var i1 = 0
+    var i2 = 0
+    var i3 = 0
+    while (!hasFinished) {
+      while (i1 < largestSize1 && i2 < largestSize2) {
+        func(
+          tensor1Data, tensor1Offset + i1 * tensor1Stride,
+          tensor2Data, tensor2Offset + i2 * tensor2Stride,
+          tensor3Data, tensor3Offset + i3 * tensor3Stride
+        )
+        i1 = i1 + 1
+        i2 = i2 + 1
+        i3 = i3 + 1
+      }
+
+      if (i1 == largestSize1) {
+        val r = updateCounter(tensor1, counter1, tensor1Offset, largestDim1)
+        hasFinished = r._1
+        tensor1Offset = r._2
+        i1 = 0
+      }
+
+      if (i2 == largestSize2) {
+        val r = updateCounter(tensor2, counter2, tensor2Offset, largestDim2)
+        hasFinished = r._1
+        tensor2Offset = r._2
+        i2 = 0
+      }
+
+      if (i3 == largestSize3) {
+        val r = updateCounter(tensor3, counter3, tensor3Offset, largestDim3)
+        hasFinished = r._1
+        tensor2Offset = r._2
+        i3 = 0
+      }
+    }
+  }
+
 }
